Retrieve rendered mighty code component

Retrieves and renders a mighty code component, supporting both built-in standards and custom components from local files.

Usage

get_rendered_component(code_component, params = NULL)

Arguments

code_component

Character string specifying either a standard component name or path to a custom component file (R or Mustache template)

params

named list of input parameters. Passed along to mighty_component$render()

Value

An object of class mighty_component_rendered containing the rendered code template

Details

Processes different component types based on file extension or component name. For R files, extracts and renders custom functions. For Mustache templates, creates components from template files. For standard names, retrieves and renders built-in standard components.

See also

Examples

get_rendered_component("ady", list(variable = "ASTDY", date = "ASTDT"))
#> <mighty_component_rendered/mighty_component/R6>
#> Type: derivation
#> Depends:
#> • .self.ASTDT
#> • .self.TRTSDT
#> Outputs:
#> • ASTDY
#> Code:
#> .self <- .self |>
#>   dplyr::mutate(
#>     ASTDY = admiral::compute_duration(
#>       start_date = TRTSDT,
#>       end_date = ASTDT,
#>       in_unit = "days",
#>       out_unit = "days",
#>       add_one = TRUE
#>     )
#>   )