Retrieve mighty standard component

Retrieve either the generalised standard component (template) or the rendered standard component with code that is ready to use.

get_standard(): Returns an object of class mighty_standard

get_rendered_standard(): Returns an object of class mighty_standard_rendered

When rendering the standard the required list of parameters depends on the standard. Check the documentation of the specific standard for details.

Usage

get_standard(standard)

get_rendered_standard(standard, params)

Arguments

standard

character name of the standard component to retrieve.

params

named list of input parameters. Passed along to mighty_standard$render().

See also

Examples

get_standard("ady")
#> <mighty_standard>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     code: active binding
#>     depends: active binding
#>     document: function () 
#>     initialize: function (template) 
#>     outputs: active binding
#>     render: function (...) 
#>     template: active binding
#>     type: active binding
#>   Private:
#>     .code:  .self <- .self |>   dplyr::mutate(     {{ variable }} = ...
#>     .depends:  .self {{ date }}  .self TRTSDT
#>     .outputs:  {{ variable }}
#>     .template: #' Analysis relative day #' @description #' Derives the  ...
#>     .type:  derivation

get_rendered_standard("ady", list(variable = "ASTDY", date = "ASTDT"))
#> <mighty_standard_rendered>
#>   Inherits from: <mighty_standard>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     code: active binding
#>     coverage: function () 
#>     depends: active binding
#>     document: function () 
#>     eval: function (envir = parent.frame()) 
#>     initialize: function (template) 
#>     outputs: active binding
#>     render: function (...) 
#>     stream: function (path) 
#>     template: active binding
#>     test: function (input, expected) 
#>     type: active binding
#>   Private:
#>     .code:  .self <- .self |>   dplyr::mutate(     ASTDY = admiral: ...
#>     .depends:  .self ASTDT  .self TRTSDT
#>     .outputs:  ASTDY
#>     .template: #' Analysis relative day #' @description #' Derives the  ...
#>     .type:  derivation